+++ client/v2/CHANGELOG.md
@@ -34,11 +34,11 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
-## Unreleased
+## UNRELEASED
 
 ### Bug Fixes
 
-* (autocli) #26673 Correctly encode `cosmos.Dec` fields supplied in positional JSON arguments.
+* (autocli) [#26735](https://github.com/cosmos/cosmos-sdk/pull/26735) Correctly encode `cosmos.Dec` fields supplied in positional JSON arguments.
 
 ## [v2.0.0-beta.11](https://github.com/cosmos/cosmos-sdk/tree/client/v2.0.0-beta.11) - 2025-05-30